Make Intel Vermilion cards use the vesa driver
The vermilion driver is not maintained, doesn't compile and doesn't work since a
long time ago. There are no bug reports for it.
Vermilion definition inside the Cards+ file was not dropped because we need to
keep documented that those pci IDs are Vermilion cards, and not any other intel
video card.


git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/ldetect-lst/trunk@266269 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94